pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mldlc-governance-framework"
version = "0.1.0"
description = "Machine Learning Development Lifecycle Governance Framework"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "streamlit>=1.28",
    "pandas>=2.0",
    "scikit-learn>=1.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "agents*", "pipeline*"]
